+++ packages/falso/src/index.ts
@@ -150,3 +150,7 @@ export { randPermission } from './lib/permission';
 export { randEmailProvider } from './lib/email-provider';
 export { randClothingSize } from './lib/clothing-size';
 export { randMovieCharacter } from './lib/movie-character';
+export { randPronoun } from './lib/pronoun';
+export { randQuote } from './lib/quote';
+export { randEmoji } from './lib/emoji';
+export { randDrinks } from './lib/drinks';